# fix: AU-1809: Oma asiointi page performance issues