--- a/public/modules/custom/grants_handler/grants_handler.module
+++ b/public/modules/custom/grants_handler/grants_handler.module
@@ -137,12 +137,15 @@ function grants_handler_theme(): array {
       'variables' => [
         'submission' => NULL,
         'document' => NULL,
+        'webform' => NULL,
+        'submission_id' => NULL,
       ],
     ],
 
     'application_status_tag' => [
       'variables' => [
         'applicationID' => NULL,
+        'application_data' => NULL,
         'langcode' => NULL,
       ],
     ],
@@ -1331,10 +1334,13 @@ function grants_handler_preprocess_application_list_item(&$variables) {
   $tOpts = ['context' => 'grants_handler'];
 
   /** @var \Drupal\webform\Entity\WebformSubmission $submission */
-  $submission = $variables['submission'];
-  $webform = $submission->getWebform();
-  $submissionData = $submission->getData();
+  $submissionData = $variables['submission'];
+  $submissionId = $variables['submission_id'];
+  $webform = $variables['webform'];
 
+  if (!$webform || !$submissionData || !$submissionId) {
+    return;
+  }
   $applicationTypes = ApplicationHandler::getApplicationTypes();
 
   $variables['applicationFormName'] = $applicationTypes[$submissionData["application_type"]][$language];
@@ -1389,7 +1395,7 @@ function grants_handler_preprocess_application_list_item(&$variables) {
     'grants_handler.edit_application',
     [
       'webform' => $webform->id(),
-      'webform_submission' => $submission->id(),
+      'webform_submission' => $submissionId,
     ],
     [
       'attributes' => [
@@ -1404,6 +1410,7 @@ function grants_handler_preprocess_application_list_item(&$variables) {
     '#theme' => 'application_status_tag',
     '#langcode' => $language,
     '#applicationID' => $submissionData["application_number"],
+    '#application_data' => $submissionData,
   ];
 
   $applicationStatuses = ApplicationHandler::getApplicationStatuses();
@@ -1474,8 +1481,12 @@ function grants_handler_preprocess_application_list_item(&$variables) {
 function grants_handler_preprocess_application_status_tag(&$variables): void {
 
   try {
-    $submissionObject = ApplicationHandler::submissionObjectFromApplicationNumber($variables['applicationID']);
-    $submissionData = $submissionObject->getData();
+    if ($variables['application_data']) {
+      $submissionData = $variables['application_data'];
+    } else {
+      $submissionObject = ApplicationHandler::submissionObjectFromApplicationNumber($variables['applicationID']);
+      $submissionData = $submissionObject->getData();
+    }
     if (array_key_exists($submissionData["status"], ApplicationHandler::getApplicationStatuses())) {
       $config = \Drupal::config('grants_handler.settings');
       $statusStrings = $config->get('statusStrings');