--- a/public/modules/custom/grants_handler/src/ApplicationHandler.php
+++ b/public/modules/custom/grants_handler/src/ApplicationHandler.php
@@ -4,6 +4,7 @@ namespace Drupal\grants_handler;
 
 use Drupal\Component\Serialization\Json;
 use Drupal\Core\Database\Connection;
+use Drupal\Core\Database\Database;
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\Core\Language\LanguageManager;
 use Drupal\Core\Logger\LoggerChannel;
@@ -400,7 +401,7 @@ class ApplicationHandler {
   /**
    * Check if given submission is allowed to be edited.
    *
-   * @param \Drupal\webform\Entity\WebformSubmission|null $submission
+   * @param array|null $submission
    *   Submission in question.
    * @param string $status
    *   If no object is available, do text comparison.
@@ -408,13 +409,13 @@ class ApplicationHandler {
    * @return bool
    *   Is submission editable?
    */
-  public static function isSubmissionFinished(?WebformSubmission $submission, string $status = ''): bool {
+  public static function isSubmissionFinished($submission, string $status = ''): bool {
     if (NULL === $submission) {
       $submissionStatus = $status;
     }
     else {
-      $data = $submission->getData();
-      $submissionStatus = $data['status'];
+      #$data = $submission->getData();
+      $submissionStatus = $submission['status'];
     }
 
     $applicationStatuses = self::getApplicationStatuses();
@@ -1735,6 +1736,190 @@ class ApplicationHandler {
    * @throws \GuzzleHttp\Exception\GuzzleException
    * @throws \Drupal\helfi_helsinki_profiili\TokenExpiredException
    */
+  public static function getCompanyApplicationsListing(
+    array $selectedCompany,
+    string $appEnv,
+    bool $sortByFinished = FALSE,
+    bool $sortByStatus = FALSE,
+    string $themeHook = ''): array {
+
+    /** @var \Drupal\helfi_atv\AtvService $atvService */
+    $atvService = \Drupal::service('helfi_atv.atv_service');
+
+    /** @var \Drupal\grants_profile\GrantsProfileService $grantsProfileService */
+    $grantsProfileService = \Drupal::service('grants_profile.service');
+
+    /** @var \Drupal\helfi_helsinki_profiili\HelsinkiProfiiliUserData $helsinkiProfiiliService */
+    $helsinkiProfiiliService = \Drupal::service('helfi_helsinki_profiili.userdata');
+    $userData = $helsinkiProfiiliService->getUserData();
+
+    $applications = [];
+    $finished = [];
+    $unfinished = [];
+
+    $selectedRoleData = $grantsProfileService->getSelectedRoleData();
+
+    $lookForAppEnv = 'appenv:' . $appEnv;
+
+    if ($selectedRoleData['type'] == 'private_person') {
+      $searchParams = [
+        'service' => 'AvustushakemusIntegraatio',
+        'user_id' => $userData['sub'],
+        'lookfor' => $lookForAppEnv . ',applicant_type:' . $selectedRoleData['type'],
+      ];
+    }
+    elseif ($selectedRoleData['type'] == 'unregistered_community') {
+      $searchParams = [
+        'service' => 'AvustushakemusIntegraatio',
+        'user_id' => $userData['sub'],
+        'lookfor' => $lookForAppEnv . ',applicant_type:' . $selectedRoleData['type'] .
+          ',applicant_id:' . $selectedRoleData['identifier'],
+      ];
+    }
+    else {
+      $searchParams = [
+        'service' => 'AvustushakemusIntegraatio',
+        'business_id' => $selectedCompany['identifier'],
+        'lookfor' => $lookForAppEnv . ',applicant_type:' . $selectedRoleData['type'],
+      ];
+    }
+
+    $applicationDocuments = $atvService->searchDocuments($searchParams);
+
+    /** @var \Drupal\grants_metadata\AtvSchema $atvSchema */
+    $atvSchema = \Drupal::service('grants_metadata.atv_schema');
+
+    /**
+     * Create rows for table.
+     *
+     * @var  \Drupal\helfi_atv\AtvDocument $document
+     */
+    foreach ($applicationDocuments as $document) {
+      // Make sure the type is acceptable one.
+      $docArray = $document->toArray();
+      $id = AtvSchema::extractDataForWebForm(
+        $docArray['content'], ['applicationNumber']
+      );
+
+      if (!isset($id['applicationNumber']) || empty($id['applicationNumber'])) {
+        continue;
+      }
+      if (array_key_exists($document->getType(), ApplicationHandler::getApplicationTypes())) {
+
+        try {
+          $dataDefinition = self::getDataDefinition($document->getType());
+          $convertedData = $atvSchema->documentContentToTypedData(
+            $document->getContent(),
+            $dataDefinition,
+            $document->getMetadata()
+          );
+          $applicationNumber = $convertedData['application_number'];
+          $submissionSerial = self::getSerialFromApplicationNumber($applicationNumber);
+          $webform = self::getWebformFromApplicationNumber($applicationNumber);
+
+          if (!$webform || !$submissionSerial) {
+            continue;
+          }
+
+          $database = Database::getConnection();
+          $query = $database->select('webform_submission', 'ws')
+            ->fields('ws', ['sid'])
+            ->condition('ws.serial', $submissionSerial)
+            ->condition('ws.webform_id', $webform->id());
+          $result = $query->execute();
+          $sid = $result->fetchField();
+          $convertedData['messages'] = self::parseMessages($convertedData);
+        }
+        catch (\Throwable $e) {
+          \Drupal::logger('application_handler')->error(
+            'Failed to get submission object from application number. Submission skipped in application listing. ID: @id Error: @error',
+            [
+              '@error' => $e->getMessage(),
+              '@id'    => $document->getTransactionId(),
+            ]
+          );
+          continue;
+        }
+        $ts = strtotime($convertedData['form_timestamp_created'] ?? '');
+        if ($themeHook !== '') {
+          $submission = [
+            '#theme' => $themeHook,
+            '#submission' => $convertedData,
+            '#document' => $document,
+            '#webform' => $webform,
+            '#submission_id' => $sid,
+          ];
+        }
+        else {
+          $submission = $convertedData;
+        }
+        if ($sortByFinished === TRUE) {
+          if (self::isSubmissionFinished($convertedData)) {
+            $finished[$ts] = $submission;
+          }
+          else {
+            $unfinished[$ts] = $submission;
+          }
+        }
+        elseif ($sortByStatus === TRUE) {
+          $applications[$convertedData['status']][$ts] = $submission;
+        }
+        else {
+          $applications[$ts] = $submission;
+        }
+      }
+    }
+    if ($sortByFinished === TRUE) {
+      ksort($finished);
+      ksort($unfinished);
+      return [
+        'finished' => $finished,
+        'unifinished' => $unfinished,
+      ];
+    }
+    elseif ($sortByStatus === TRUE) {
+      $applicationsSorted = [];
+      foreach ($applications as $key => $value) {
+        krsort($value);
+        $applicationsSorted[$key] = $value;
+      }
+      ksort($applicationsSorted);
+      return $applicationsSorted;
+    }
+    else {
+      ksort($applications);
+      return $applications;
+    }
+  }
+
+  /**
+   * Get company applications, either sorted by finished or all in one array.
+   *
+   * @param array $selectedCompany
+   *   Company data.
+   * @param string $appEnv
+   *   Environment.
+   * @param bool $sortByFinished
+   *   When true, results will be sorted by finished status.
+   * @param bool $sortByStatus
+   *   Sort by application status.
+   * @param string $themeHook
+   *   Use theme hook to render content. Set this to theme hook wanted to use,
+   *   and sen #submission to webform submission.
+   *
+   * @return array
+   *   Submissions in array.
+   *
+   * @throws \Drupal\Component\Plugin\Exception\InvalidPluginDefinitionException
+   * @throws \Drupal\Component\Plugin\Exception\PluginNotFoundException
+   * @throws \Drupal\Core\Entity\EntityStorageException
+   * @throws \Drupal\Core\TempStore\TempStoreException
+   * @throws \Drupal\grants_mandate\CompanySelectException
+   * @throws \Drupal\helfi_atv\AtvDocumentNotFoundException
+   * @throws \Drupal\helfi_atv\AtvFailedToConnectException
+   * @throws \GuzzleHttp\Exception\GuzzleException
+   * @throws \Drupal\helfi_helsinki_profiili\TokenExpiredException
+   */
   public static function getCompanyApplications(
     array $selectedCompany,
     string $appEnv,
@@ -1800,7 +1985,6 @@ class ApplicationHandler {
       if (!isset($id['applicationNumber']) || empty($id['applicationNumber'])) {
         continue;
       }
-
       if (array_key_exists($document->getType(), ApplicationHandler::getApplicationTypes())) {
         try {
           $submissionObject = self::submissionObjectFromApplicationNumber($document->getTransactionId(), $document);