--- a/public/modules/custom/grants_handler/src/ApplicationHandler.php
+++ b/public/modules/custom/grants_handler/src/ApplicationHandler.php
@@ -1819,7 +1819,7 @@ class ApplicationHandler {
           continue;
         }
 
-        if (!$submissionData) {
+        if (!$submissionData || !$submissionId) {
           continue;
         }