--- a/public/modules/custom/grants_handler/templates/application-list.html.twig
+++ b/public/modules/custom/grants_handler/templates/application-list.html.twig
@@ -51,5 +51,10 @@
         {{ item }}
       {% endfor %}
     </ul>
+    {% if type != 'drafts' %}
+      <div class="hds-pagination-container">
+        <ul class="pagination hds-pagination"></ul>
+      </div>
+    {% endif %}
   </div>
 </div>