--- a/public/modules/custom/grants_handler/templates/application-list.html.twig
+++ b/public/modules/custom/grants_handler/templates/application-list.html.twig
@@ -53,7 +53,7 @@
     </ul>
     {% if type != 'drafts' %}
       <div class="hds-pagination-container">
-        <ul class="pagination hds-pagination"></ul>
+        <ul class="pagination application-list__pagination"></ul>
       </div>
     {% endif %}
   </div>