--- a/public/modules/custom/grants_oma_asiointi/src/Controller/GrantsOmaAsiointiController.php
+++ b/public/modules/custom/grants_oma_asiointi/src/Controller/GrantsOmaAsiointiController.php
@@ -110,6 +110,7 @@ class GrantsOmaAsiointiController extends ControllerBase implements ContainerInj
    *   Render array
    */
   public function build(): array {
+
     $selectedCompany = $this->grantsProfileService->getSelectedRoleData();
 
     if ($selectedCompany == NULL) {
@@ -131,7 +132,7 @@ class GrantsOmaAsiointiController extends ControllerBase implements ContainerInj
 
     try {
       // Get applications from ATV.
-      $applications = ApplicationHandler::getCompanyApplications(
+      $applications = ApplicationHandler::getCompanyApplicationsListing(
         $selectedCompany,
         $appEnv,
         FALSE,
@@ -152,7 +153,7 @@ class GrantsOmaAsiointiController extends ControllerBase implements ContainerInj
     foreach ($applications as $values) {
       $other = array_merge($other, $values);
       foreach ($values as $application) {
-        $appMessages = ApplicationHandler::parseMessages($application['#submission']->getData());
+        $appMessages = ApplicationHandler::parseMessages($application['#submission']);
         foreach ($appMessages as $msg) {
           if ($msg["messageStatus"] == 'UNREAD' && $msg["sentBy"] == 'Avustusten kasittelyjarjestelma') {
             $unreadMsg[] = [