--- a/public/modules/custom/grants_oma_asiointi/src/Controller/GrantsOmaAsiointiController.php
+++ b/public/modules/custom/grants_oma_asiointi/src/Controller/GrantsOmaAsiointiController.php
@@ -132,7 +132,7 @@ class GrantsOmaAsiointiController extends ControllerBase implements ContainerInj
 
     try {
       // Get applications from ATV.
-      $applications = ApplicationHandler::getCompanyApplicationsListing(
+      $applications = ApplicationHandler::getCompanyApplications(
         $selectedCompany,
         $appEnv,
         FALSE,