--- a/public/themes/custom/hdbt_subtheme/src/scss/components/grants-profile/_application-list.scss
+++ b/public/themes/custom/hdbt_subtheme/src/scss/components/grants-profile/_application-list.scss
@@ -209,4 +209,54 @@
       white-space: nowrap;
     }
   }
+  &__pagination {
+    @include font('body');
+    font-weight: $font-weight-bold;
+    display: inline-flex;
+    flex-wrap: wrap;
+    gap: var(--spacing-2-xs);
+    justify-content: center;
+    list-style: none;
+    padding: 0;
+
+    li {
+      margin-top: 0;
+
+      &.active a {
+        text-decoration: underline;
+      }
+
+      &.disabled a {
+        text-decoration: none;
+        cursor: default;
+
+        &:hover,
+        &:focus,
+        &:active {
+          text-decoration: none;
+          outline: none;
+        }
+      }
+
+      a {
+        display: block;
+        color: var(--color-black-90);
+        height: var(--link-height);
+        min-width: 40px;
+        padding-left: var(--spacing-3-xs);
+        padding-right: var(--spacing-3-xs);
+        text-decoration: none;
+
+        &:hover {
+          text-decoration: underline;
+        }
+
+        &:focus,
+        &:active {
+          outline: solid var(--color-coat-of-arms) 3px;
+          outline-offset: 1px;
+        }
+      }
+    }
+  }
 }