… public/modules/custom/grants_oma_asiointi/src/Controller/ApplicationsListController.php
… public/themes/custom/hdbt_subtheme/dist/css/styles.min.css